`UPD dashboard & fix paiment & fix pointage`

+++ Infrastructures/Domains/Models/Dashboard/DashboardPointage.cs
@@ -12,7 +12,7 @@ namespace Infrastructures.Domains.Models.Dashboard
         public string NomComplet { get; set; }
         public int NombrePresences { get; set; }
         public int NombreAbsences { get; set; }
-        public int NombreHeuresSupp { get; set; }
+        public decimal NombreHeuresSupp { get; set; }
         public TimeSpan? EntryHeure { get; set; }
         public TimeSpan? ExitHeure { get; set; }
