+++ Infrastructures/Domains/Models/Employee/Employe.cs
@@ -17,7 +17,7 @@
         public DateTime DateEntree { get; set; }
         public DateTime? DateSortie { get; set; }
         public byte[] Photo { get; set; }
-        public int Journee { get; set; }
+        public decimal Journee { get; set; }
         public string FonctionName { get; set; }
         public string Status { get; set; }
         public string FullName => $"{Nom} {Prenom}";
